Rename s3Client to awsClient and pull in SES code
